% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetstats-package.R
\docType{package}
\name{rtweetstats-package}
\alias{rtweetstats}
\alias{rtweetstats-package}
\title{rtweetstats: Functions to Facilitate Data Science Projects}
\description{
This package is compiled of two functions: keystats() and 
    userstats(), both of which return HTML outputs with summary information based
    on user input.
        Upon first use of the functions, the user will be asked to authorize a Twitter
    API token that will continue to be valid and functional following first use. As
    such, these functions assume that the user has access to a Twitter account of
    their own.
        keystats() allows Twitter users to search the website by keyword or hashtag,
    returning multiple graphics and statistical summaries such as a bar plot of
    related unique words, a word cloud, a radar chart, and two bar plots of location
    statistics in an HTML output.
        userstats() allows Twitter users to search the website by username, allowing the 
    user to search multiple usernames at one time, specify which year they would like 
    summary information for, specify the number of tweets they would like to analyze 
    (with a maximum output allowed of 3,200 tweets), customize the background color 
    of output tables, and customize the text color of the output with the assumption 
    of knowledge of HTML color codes. The output is returned as HTML.
        Please see the help files of each respective function for documentation of 
    how they may be used and navigated.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/annasophiepeters/rtweetstats/issues}
}

}
\author{
\strong{Maintainer}: Anna Sophie Peters \email{aspeters@wesleyan.edu}

Authors:
\itemize{
  \item Jiner Zheng \email{jzheng@wesleyan.edu}
  \item Nurah Jaradat \email{njaradat@wesleyan.edu}
}

}
\keyword{internal}
